Changing -(z-1) to (1-z) and (z-1)^2 to (1-z)^2
The calculations show results with (z-1) instead of (1-z) and there is a reference to the similarity to Bernstein's polynomials. As a matter of fact, with simple sign manipulation it can be shown that the intermediate control points are actually always the result of the application of the Bernstein polynomials (of course using z instead of t).